requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 文档处理
python-docx==1.1.0
reportlab==4.0.7
openpyxl==3.1.2
PyPDF2==3.0.1
unstructured==0.11.6
unstructured[pdf]==0.11.6
unstructured[docx]==0.11.6

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 日期时间处理
python-dateutil==2.8.2

# 环境变量
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 缓存
redis==5.0.1

# 文件处理
aiofiles==23.2.1
Pillow==10.1.0

# 模板引擎
jinja2==3.1.2

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# 其他工具
click==8.1.7
rich==13.7.0
typer==0.9.0